Made massive changes to the android manifest. Please make sure to check and make sure it matches yours.
After making changes to the manifest, uninstall the current versions you have on your emulators and redeploy the app so it reinstalls.

All other changes are in SubmitIdeaFragment.CS

I have made a lot of comments to explain changes. LMK if it makes sense

Lastly lol, I need help with the SendEmail method. Am not used to working with fragments and all the examples on the web explain using intents.

Almost there guys.